Переносит разметку в классы
