+++ src/components/film-content/films-list.js
@@ -1,0 +1,29 @@
+import {createElement} from "../../util";
+
+// Генерация Контейнера списка фильмов
+export default class FilmsListComponent {
+  constructor() {
+    this._element = null;
+  }
+
+  getTemplate() {
+    return (
+      `<section class="films-list">
+      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+      <div class="films-list__container">
+     </div>
+      </section>`);
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
