+++ src/components/header/profile.js
@@ -1,22 +1,18 @@
-import {createElement} from "../util";
+import {createElement} from "../../util";
 
 // Генерация Звания пользователя
-const createProfileTemplate = () => {
-  return (
-    `<section class="header__profile profile">
-    <p class="profile__rating">Movie Buff</p>
-    <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-  </section>`
-  );
-};
-
 export default class ProfileComponent {
   constructor() {
     this._element = null;
   }
 
   getTemplate() {
-    return createProfileTemplate();
+    return (
+      `<section class="header__profile profile">
+      <p class="profile__rating">Movie Buff</p>
+      <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
+    </section>`
+    );
   }
 
   getElement() {
