--- src/components/sort.js
@@ -1,34 +1,0 @@
-import {createElement} from "../util";
-
-// Генерерация Сортировки
-const createSortTemplate = () => {
-  return (
-    `<ul class="sort">
-    <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
-    <li><a href="#" class="sort__button">Sort by date</a></li>
-    <li><a href="#" class="sort__button">Sort by rating</a></li>
-  </ul>`
-  );
-};
-
-export default class SortComponent {
-  constructor() {
-    this._element = null;
-  }
-
-  getTemplate() {
-    return createSortTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
-}
