+++ src/main.js
@@ -1,12 +1,12 @@
-import ProfileComponent from './components/profile';
-import FiltersComponent from "./components/menu";
-import SortComponent from "./components/sort";
-import FilmContentComponent from "./components/film-content";
-import FilmsListComponent from "./components/films-list";
-import FilmCardComponent from "./components/film-card";
-import ShowMoreButtonComponent from "./components/show-more-button";
-import FilmTopRatedComponent from "./components/film-extra-rated";
-import FilmMostCommentedComponent from "./components/film-extra-commented";
+import ProfileComponent from './components/header/profile';
+import FiltersComponent from "./components/header/menu";
+import SortComponent from "./components/header/sort";
+import FilmContentComponent from "./components/film-content/film-content";
+import FilmsListComponent from "./components/film-content/films-list";
+import FilmCardComponent from "./components/film-content/film-card";
+import ShowMoreButtonComponent from "./components/film-content/show-more-button";
+import FilmTopRatedComponent from "./components/film-content/film-extra-rated";
+import FilmMostCommentedComponent from "./components/film-content/film-extra-commented";
 import FilmCountComponent from "./components/statistic";
 import {generateCards} from "./mock/film-card";
 import {generateFilters} from "./mock/filter";
